requirements.txt
Django~=5.1.1
djangorestframework-simplejwt~=5.3.1
django-filter~=1.1.0
drf-spectacular~=0.27.2
djangorestframework~=3.15.2
twilio~=9.3.2